Optional Manifest file
