+++ tests/Controller/DummyController.php
@@ -12,6 +12,16 @@ use Symfony\Component\Routing\Attribute\Route;
  */
 final class DummyController extends AbstractController
 {
+    #[Route('/privacy-policy', name: 'privacy_policy')]
+    public function privacyPolicy(string $param1): void
+    {
+    }
+
+    #[Route('/terms-of-service', name: 'terms_of_service')]
+    public function tos(string $param1): void
+    {
+    }
+
     #[Route('/audio-file-handler/{param1}', name: 'audio_file_handler')]
     public function dummy1(string $param1): void
     {
