+++ tests/Functional/DevServerTest.php
@@ -38,5 +38,6 @@ final class DevServerTest extends WebTestCase
         // Then
         static::assertResponseIsSuccessful();
         static::assertResponseHeaderSame('Content-Type', 'application/javascript');
+        dump($client->getResponse()->getContent());
     }
 }
